Set HTTP headers as an associative array
MultiHttpClient expects headers in an associative array. The extension instead
sent headers as an array of strings of the form 'Cache-control: no-cache', which
silently produced HTTP headers of the form '0: Cache-control: no-cache'.
Needless to say, that rendered those headers useless, which in turn caused
template updates to not happen as the no-cache header was never seen. The same
issue also affected revision deletion updates.

Also: Fix the default server port to use RESTBase's default port.

Bug: T92703
Change-Id: I57082560cb6b00d1bbe05b8f06c686475803f160